pyproject.toml
[tool.poetry]
authors = ["m09 <142691+m09@users.noreply.github.com>"]
description = ""
name = "ghcli"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
attrs = "^20.3.0"
deal = "^4.5.0"
python = "^3.8"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-bugbear = "^20.11.1"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
mutmut = "^2.1.0"
mypy = "^0.812"
pylint = "^2.7.2"
pytest = "^5.2"

[tool.poetry.scripts]
ghcli = 'ghcli.cli:main'

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
